+++ sdk/containerservice/arm-containerservice/src/operationsInterfaces/index.ts
@@ -19,3 +19,4 @@ export * from "./snapshots";
 export * from "./managedClusterSnapshots";
 export * from "./trustedAccessRoles";
 export * from "./trustedAccessRoleBindings";
+export * from "./loadBalancers";